자동로또와 수동로또 클래스 분리, LottoCount 객체 생성
