+++ src/main/java/step4/model/LottoCount.java
@@ -1,0 +1,14 @@
+package step4.model;
+
+public class LottoCount {
+
+    private final int count;
+
+    public LottoCount(int count) {
+        this.count = count;
+    }
+
+    public int getCount() {
+        return this.count;
+    }
+}
